feat: Add new utility tools and achieve 100% test coverage in tools.py
- Add `json_validator`, `hash_generator`, and `url_parser` tools.
- Implement corresponding handlers in `backend/api/tools.py`.
- Add test cases in `backend/tests/test_tools_api.py` covering new and previously untested functions.
- Achieve 100% test coverage for `api.tools`.
- Handle potential None input parameters gracefully in newly added handlers.
- Clean up redundant code and duplicate blocks from previous merge attempts.
