+++ .github/workflows/pr-governance.yml
@@ -25,8 +25,8 @@ permissions:
   contents: read
   checks: read
   statuses: read
-  pull-requests: read
-  issues: read
+  pull-requests: write
+  issues: write
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.event.workflow_run.pull_requests[0].number || github.event.check_run.pull_requests[0].number || github.event.inputs.pr_number || github.run_id }}
@@ -34,14 +34,8 @@ concurrency:
 
 jobs:
   governance:
-    name: PR governance metadata controller
+    name: metadata-only gate evaluation
     runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      checks: write
-      statuses: read
-      pull-requests: write
-      issues: write
     if: >-
       ${{
         github.event_name == 'pull_request_target' ||
