--- backend/requirements-agent.txt
@@ -1,40 +1,0 @@
-# Optional dependencies for the Noema general agent (services/noema_agent.py).
-#
-# Pydantic-AI is MIT licensed and builds on the OpenAI-compatible provider layer
-# already used by naruon (openai, httpx, pydantic). Installing this enables the
-# agent; without it, services/noema_agent.py degrades to a structured no-op.
-#
-# Kept in a small hash-pinned optional set so the core runtime lock remains
-# focused while CI can still run the real agent build-path test deterministically.
-# Install together with the core lock file:
-#     pip install --require-hashes -r backend/requirements-hashes.txt -r backend/requirements-agent.txt
-#
-# Pinned to the pydantic-ai 2.x line: services/noema_agent.py imports
-# ``OpenAIChatModel`` from ``pydantic_ai.models.openai`` (renamed from
-# ``OpenAIModel`` in 2.x) and constructs the model via ``OpenAIProvider``.
-# Version 2.8.0 is the newest 2.x release compatible with the core
-# requirements-hashes.txt pin of openai==2.44.0; 2.9.0 requires openai>=2.45.0.
-pydantic-ai-slim[openai]==2.8.0 \
-    --hash=sha256:ddf8ac31442d5433ffff7c47af895088c230448fd28926a713c0f47aa2affc12
-    # via -r backend/requirements-agent.txt
-genai-prices==0.0.71 \
-    --hash=sha256:1d13111563af2b1ce43ccfacf77b7ac3216ad704c644408a56e11b181fe0d128
-    # via pydantic-ai-slim
-griffelib==2.1.0 \
-    --hash=sha256:cc7b3d2d2865ad0b909fcc38086e3f554b5ea7acbaa7bbb7ecaa3f5dfb7d9f00
-    # via pydantic-ai-slim
-httpcore2==2.5.0 \
-    --hash=sha256:5ce35188de461d31e8d000bfb8ef8bf22c6c16587a211e5571deaa5e9bdf842a
-    # via httpx2
-httpx2==2.5.0 \
-    --hash=sha256:3d2d4d9cf4b61f1a1f46a95947cfdb47e80cb56a2f91c6256ac8f58e4891df41
-    # via genai-prices
-logfire-api==4.37.0 \
-    --hash=sha256:1d756f8ba23aa56d438e0ba2c0f529a00fcac975b8785c561b058267f9465088
-    # via pydantic-graph
-pydantic-graph==2.8.0 \
-    --hash=sha256:9b44a06d413586299a2d9ffb76efc1aa1d3b2268984332b03af24dd0c32dad35
-    # via pydantic-ai-slim
-truststore==0.10.4 \
-    --hash=sha256:adaeaecf1cbb5f4de3b1959b42d41f6fab57b2b1666adb59e89cb0b53361d981
-    # via httpcore2, httpx2
