+++ backend/services/project_graph/__init__.py
@@ -18,8 +18,6 @@ from .project_registration import (
     ProjectCandidateSummary,
     ProjectCitation,
     ProjectCorrection,
-    ProjectDecisionRecord,
-    ProjectDecisionView,
     ProjectEvidence,
     ProjectGraphNotFoundError,
     ProjectGraphQueryScope,
@@ -32,7 +30,6 @@ from .project_registration import (
     ProjectTraceability,
     apply_project_correction,
     confirm_project_candidate,
-    get_project_decisions,
     get_project_evidence,
     get_project_relation_summary,
     get_project_traceability,
@@ -47,8 +44,6 @@ __all__ = [
     "ProjectCandidateSummary",
     "ProjectCitation",
     "ProjectCorrection",
-    "ProjectDecisionRecord",
-    "ProjectDecisionView",
     "ProjectEvidence",
     "ProjectGraphNotFoundError",
     "ProjectGraphQueryScope",
@@ -69,7 +64,6 @@ __all__ = [
     "apply_project_graph_correction",
     "confirm_project_candidate",
     "extract_project_semantics",
-    "get_project_decisions",
     "get_project_evidence",
     "get_project_relation_summary",
     "get_project_traceability",
