+++ backend/services/project_graph/traceability.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 from .project_registration import (
-    ProjectDecisionRecord,
-    ProjectDecisionView,
     ProjectEvidence,
     ProjectRelationSummary,
     ProjectRelationTypeSummary,
@@ -11,15 +9,12 @@ from .project_registration import (
     ProjectTraceRelation,
     ProjectTraceRelationEndpoint,
     ProjectTraceability,
-    get_project_decisions,
     get_project_evidence,
     get_project_relation_summary,
     get_project_traceability,
 )
 
 __all__ = [
-    "ProjectDecisionRecord",
-    "ProjectDecisionView",
     "ProjectEvidence",
     "ProjectRelationSummary",
     "ProjectRelationTypeSummary",
@@ -28,7 +23,6 @@ __all__ = [
     "ProjectTraceRelation",
     "ProjectTraceRelationEndpoint",
     "ProjectTraceability",
-    "get_project_decisions",
     "get_project_evidence",
     "get_project_relation_summary",
     "get_project_traceability",
