+++ backend/tests/test_project_graph_llm_extractor.py
@@ -416,236 +416,6 @@ async def test_self_loop_and_duplicate_relations_are_dropped(monkeypatch):
     assert relation_edges[0].edge_type == "implements"
 
 
-def _two_grounded_decisions() -> list["llm_extractor.ExtractedObjectPayload"]:
-    # A superseding decision (seg1) and the prior decision it replaces (seg2),
-    # each grounded in its own cited segment so a supersedes relation between
-    # them is evidenced by the union of both citations.
-    return [
-        _object(
-            object_type="decision",
-            title="Adopt Stripe as the payment gateway",
-            summary="The committee now standardizes on Stripe.",
-            segment_uids=["seg1"],
-            confidence=0.86,
-            local_key="dec-new",
-        ),
-        _object(
-            object_type="decision",
-            title="Adopt PayPal as the payment gateway",
-            summary="The earlier decision had chosen PayPal.",
-            segment_uids=["seg2"],
-            confidence=0.71,
-            local_key="dec-old",
-        ),
-    ]
-
-
-def _two_decision_segments() -> list[ProjectSourceSegment]:
-    return [
-        _segment("seg1", "We now standardize on Stripe, replacing the prior choice."),
-        _segment("seg2", "The earlier decision had chosen PayPal as the gateway."),
-    ]
-
-
-def test_decision_centric_relations_are_in_controlled_vocabulary():
-    # The DECISION entity (#1058) and the decision read model (#1061) speak in
-    # terms of a decision resolving an issue, a requirement being decided by a
-    # decision, and a decision superseding a prior one. Those relation tokens
-    # must live in the controlled vocabulary or the grounded LLM extractor would
-    # silently drop every decision-centric relation at the vocabulary gate,
-    # leaving decisions connectable only through the generic ``relates_to``.
-    assert {"resolves", "decided_by", "supersedes"} <= (
-        llm_extractor.ALLOWED_RELATION_TYPES
-    )
-
-
-@pytest.mark.asyncio
-async def test_resolves_relation_links_decision_to_issue(monkeypatch):
-    objects = [
-        _object(
-            object_type="decision",
-            title="Waive the staging gate for the hotfix",
-            summary="The lead approved shipping the hotfix without staging.",
-            segment_uids=["seg1"],
-            confidence=0.83,
-            local_key="dec-a",
-        ),
-        _object(
-            object_type="issue",
-            title="Staging environment is down",
-            summary="Staging outage blocks the release.",
-            segment_uids=["seg2"],
-            confidence=0.9,
-            local_key="iss-b",
-        ),
-    ]
-    monkeypatch.setattr(
-        llm_extractor,
-        "_call_llm",
-        AsyncMock(
-            return_value=llm_extractor.ExtractionPayload(
-                objects=objects,
-                relations=[
-                    _relation(
-                        source_local_key="dec-a",
-                        target_local_key="iss-b",
-                        relation_type="resolves",
-                        confidence=0.68,
-                    )
-                ],
-            )
-        ),
-    )
-
-    result = await llm_extractor.extract_project_semantics_llm(
-        [
-            _segment("seg1", "The lead approved shipping the hotfix without staging."),
-            _segment("seg2", "Staging outage blocks the release."),
-        ],
-        api_key="key",
-        model="gpt-test",
-    )
-
-    decision = next(o for o in result.objects if o.object_type.value == "decision")
-    issue = next(o for o in result.objects if o.object_type.value == "issue")
-    resolves_edges = [edge for edge in result.edges if edge.edge_type == "resolves"]
-    assert len(resolves_edges) == 1
-    edge = resolves_edges[0]
-    # Directionality is preserved: the decision resolves the issue, not vice
-    # versa.
-    assert edge.source_uid == decision.uid
-    assert edge.target_uid == issue.uid
-    # Grounded in the union of both endpoints' cited segments — never an uncited
-    # reference.
-    assert set(edge.source_segment_uids) == {"seg1", "seg2"}
-
-
-@pytest.mark.asyncio
-async def test_decided_by_relation_preserves_direction(monkeypatch):
-    objects = [
-        _object(
-            object_type="requirement",
-            title="Payments must use a single gateway",
-            summary="One gateway is required for reconciliation.",
-            segment_uids=["seg2"],
-            confidence=0.88,
-            local_key="req-a",
-        ),
-        _object(
-            object_type="decision",
-            title="Adopt Stripe as the payment gateway",
-            summary="The committee standardizes on Stripe.",
-            segment_uids=["seg1"],
-            confidence=0.86,
-            local_key="dec-b",
-        ),
-    ]
-    monkeypatch.setattr(
-        llm_extractor,
-        "_call_llm",
-        AsyncMock(
-            return_value=llm_extractor.ExtractionPayload(
-                objects=objects,
-                relations=[
-                    _relation(
-                        source_local_key="req-a",
-                        target_local_key="dec-b",
-                        relation_type="decided_by",
-                        confidence=0.7,
-                    )
-                ],
-            )
-        ),
-    )
-
-    result = await llm_extractor.extract_project_semantics_llm(
-        [
-            _segment("seg1", "The committee standardizes on Stripe."),
-            _segment("seg2", "One gateway is required for reconciliation."),
-        ],
-        api_key="key",
-        model="gpt-test",
-    )
-
-    requirement = next(
-        o for o in result.objects if o.object_type.value == "requirement"
-    )
-    decision = next(o for o in result.objects if o.object_type.value == "decision")
-    decided_by_edges = [
-        edge for edge in result.edges if edge.edge_type == "decided_by"
-    ]
-    assert len(decided_by_edges) == 1
-    edge = decided_by_edges[0]
-    assert edge.source_uid == requirement.uid
-    assert edge.target_uid == decision.uid
-
-
-@pytest.mark.asyncio
-async def test_supersedes_relation_links_two_decisions(monkeypatch):
-    monkeypatch.setattr(
-        llm_extractor,
-        "_call_llm",
-        AsyncMock(
-            return_value=llm_extractor.ExtractionPayload(
-                objects=_two_grounded_decisions(),
-                relations=[
-                    _relation(
-                        source_local_key="dec-new",
-                        target_local_key="dec-old",
-                        relation_type="supersedes",
-                        confidence=0.74,
-                    )
-                ],
-            )
-        ),
-    )
-
-    result = await llm_extractor.extract_project_semantics_llm(
-        _two_decision_segments(), api_key="key", model="gpt-test"
-    )
-
-    supersedes_edges = [
-        edge for edge in result.edges if edge.edge_type == "supersedes"
-    ]
-    assert len(supersedes_edges) == 1
-    edge = supersedes_edges[0]
-    assert set(edge.source_segment_uids) == {"seg1", "seg2"}
-
-
-@pytest.mark.asyncio
-async def test_decision_relation_synonym_outside_vocabulary_is_dropped(monkeypatch):
-    # Disambiguation: "overrides" is a natural-language synonym of the controlled
-    # ``supersedes`` token, but the vocabulary is a closed set of exact tokens,
-    # not a fuzzy match. A relation labelled with the synonym is dropped so the
-    # inter-object graph never accretes free-text edge labels.
-    monkeypatch.setattr(
-        llm_extractor,
-        "_call_llm",
-        AsyncMock(
-            return_value=llm_extractor.ExtractionPayload(
-                objects=_two_grounded_decisions(),
-                relations=[
-                    _relation(
-                        source_local_key="dec-new",
-                        target_local_key="dec-old",
-                        relation_type="overrides",
-                        confidence=0.9,
-                    )
-                ],
-            )
-        ),
-    )
-
-    result = await llm_extractor.extract_project_semantics_llm(
-        _two_decision_segments(), api_key="key", model="gpt-test"
-    )
-
-    assert all(
-        edge.edge_type == "segment_evidences_project_object"
-        for edge in result.edges
-    )
-
-
 # The import selector resolves through the KG extractor registry, so these
 # tests patch the extraction cores where the registry imports them.
 def _patch_extractor_cores(monkeypatch, *, llm, keyword):
