+++ connector/requirements.txt
@@ -1,1 +1,1 @@
-websockets==16.1
+websockets==16.0
