+++ frontend/package.json
@@ -9,7 +9,6 @@
     "start": "next start",
     "lint": "eslint",
     "test": "vitest run",
-    "coverage": "vitest run src/components/project-trace-readiness.test.ts --coverage --coverage.provider=v8 --coverage.reporter=json-summary --coverage.reporter=json --coverage.include=src/components/project-trace-readiness.ts",
     "typecheck": "tsc --noEmit",
     "full:smoke": "node scripts/full-product-ui-smoke.mjs",
     "pilot:smoke": "node scripts/pilot-ui-smoke.mjs",
@@ -17,12 +16,12 @@
   },
   "dependencies": {
     "@base-ui/react": "^1.6.0",
-    "@radix-ui/react-tabs": "^1.1.17",
+    "@radix-ui/react-tabs": "^1.1.15",
     "@tailwindcss/postcss": "^4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "lucide-react": "^1.24.0",
-    "next": "16.2.10",
+    "lucide-react": "^1.22.0",
+    "next": "16.2.9",
     "react": "19.2.7",
     "react-dom": "19.2.7",
     "react-resizable-panels": "^4.12.0",
@@ -37,14 +36,13 @@
     "@types/node": "^26",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "@vitest/coverage-v8": "4.1.10",
     "eslint": "^9",
-    "eslint-config-next": "16.2.10",
-    "fast-check": "^4.9.0",
+    "eslint-config-next": "16.2.9",
+    "fast-check": "^4.8.0",
     "jsdom": "^29.1.0",
     "postcss": "^8.5.16",
     "typescript": "^6",
-    "vitest": "^4.1.10"
+    "vitest": "^4.1.9"
   },
   "overrides": {
     "postcss": "^8.5.16",
