+++ registered_agents.json
@@ -1,25 +1,1 @@
-{
-  "noema-general-agent": {
-    "agent_id": "noema-general-agent",
-    "name": "Noema General Agent",
-    "framework": "pydantic-ai",
-    "framework_license": "MIT",
-    "entrypoint": "services.noema_agent:run_noema_agent",
-    "description": "General-purpose assistant that reasons over the naruon workspace (mail, content graph, tasks) on the tenant's configured LLM provider and dispatches opt-in, audit-logged writebacks through the self-hosted runner.",
-    "capabilities": [
-      "mail.search",
-      "mail.read",
-      "content_graph.query",
-      "tasks.read",
-      "tasks.update",
-      "calendar.writeback"
-    ],
-    "provider_source": "runtime_llm_provider",
-    "writeback": {
-      "opt_in": true,
-      "audit_logged": true
-    },
-    "degrades_gracefully": true,
-    "enabled": true
-  }
-}
+{}
