+++ task_agent_mapping.json
@@ -1,7 +1,1 @@
-{
-  "general": "noema-general-agent",
-  "mail.triage": "noema-general-agent",
-  "mail.search": "noema-general-agent",
-  "tasks.followup": "noema-general-agent",
-  "calendar.writeback": "noema-general-agent"
-}
+{}
